+++ tests/unit/core/test_frame.py
@@ -321,3 +321,8 @@ def test_frame__dict(dataframe):
     }
     assert sub_dataframe.row_index_lineage == ("input_1", "filter_2")
     assert sub_dataframe_dict["row_index_lineage"] == ("input_1", "filter_1")
+    loaded_sub_dataframe = Frame.parse_obj(sub_dataframe_dict)
+    # note that loaded_sub_dataframe & sub_dataframe are not fully identical (ideally, we should make them identical)
+    # loaded_sub_dataframe = input -> filter (input_frame, "MASK") -> project ["VALUE", "CUST_ID"]
+    # sub_dataframe = input -> filter (assign_frame, "MASK") -> project ["VALUE", "CUST_ID"]
+    assert sub_dataframe.preview_sql() == loaded_sub_dataframe.preview_sql()
